refactor final score map struct in pse module
