+++ x/pse/keeper/distribute.go
@@ -3,8 +3,6 @@ package keeper
 import (
 	"context"
 
-	"cosmossdk.io/collections"
-	addresscodec "cosmossdk.io/core/address"
 	sdkmath "cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
@@ -14,50 +12,19 @@ import (
 
 // DistributeCommunityPSE distributes the total community PSE amount to all delegators based on their score.
 func (k Keeper) DistributeCommunityPSE(ctx context.Context, totalPSEAmount sdkmath.Int) error {
-	var allDelegationTimeEntry []collections.KeyValue[
-		collections.Pair[sdk.ValAddress, sdk.AccAddress],
-		types.DelegationTimeEntry,
-	]
 	// iterate all delegation time entries and calculate uncalculated score.
 	var finalScoreMap = newScoreMap(k.addressCodec)
-	delegationTimeEntriesIterator, err := k.DelegationTimeEntries.Iterate(ctx, nil)
+	allDelegationTimeEntry, err := finalScoreMap.iterateDelegationTimeEntries(ctx, k)
 	if err != nil {
 		return err
-	}
-	defer delegationTimeEntriesIterator.Close()
-
-	for ; delegationTimeEntriesIterator.Valid(); delegationTimeEntriesIterator.Next() {
-		kv, err := delegationTimeEntriesIterator.KeyValue()
-		if err != nil {
-			return err
-		}
-		allDelegationTimeEntry = append(allDelegationTimeEntry, kv)
-		valAddr := kv.Key.K1()
-		delAddr := kv.Key.K2()
-		delegationTimeEntry := kv.Value
-		delegationScore, err := calculateAddedScore(ctx, k, valAddr, delegationTimeEntry)
-		if err != nil {
-			return err
-		}
-		finalScoreMap.AddScore(delAddr, delegationScore)
 	}
 
 	// add uncalculated score to account score snapshot and total score per delegator.
 	// it calculates the score from the last delegation time entry up to the current block time, which
 	// is not included in the score snapshot calculations.
-	iter, err := k.AccountScoreSnapshot.Iterate(ctx, nil)
+	err = finalScoreMap.iterateAccountScoreSnapshot(ctx, k)
 	if err != nil {
 		return err
-	}
-	defer iter.Close()
-	for ; iter.Valid(); iter.Next() {
-		kv, err := iter.KeyValue()
-		if err != nil {
-			return err
-		}
-		score := kv.Value
-		delAddr := kv.Key
-		finalScoreMap.AddScore(delAddr, score)
 	}
 
 	bondDenom, err := k.stakingKeeper.BondDenom(ctx)
@@ -116,62 +83,6 @@ func (k Keeper) DistributeCommunityPSE(ctx context.Context, totalPSEAmount sdkma
 		}
 	}
 
-	return nil
-}
-
-type scoreMap struct {
-	items []struct {
-		addr  sdk.AccAddress
-		score sdkmath.Int
-	}
-	indexMap     map[string]int
-	addressCodec addresscodec.Codec
-	totalScore   sdkmath.Int
-}
-
-func newScoreMap(addressCodec addresscodec.Codec) *scoreMap {
-	return &scoreMap{
-		items: make([]struct {
-			addr  sdk.AccAddress
-			score sdkmath.Int
-		}, 0),
-		indexMap:     make(map[string]int),
-		addressCodec: addressCodec,
-		totalScore:   sdkmath.NewInt(0),
-	}
-}
-
-func (m *scoreMap) AddScore(addr sdk.AccAddress, value sdkmath.Int) {
-	if value.IsZero() {
-		return
-	}
-	key, err := m.addressCodec.BytesToString(addr)
-	if err != nil {
-		return
-	}
-	idx, found := m.indexMap[key]
-	if !found {
-		m.items = append(m.items, struct {
-			addr  sdk.AccAddress
-			score sdkmath.Int
-		}{
-			addr:  addr,
-			score: value,
-		})
-		m.indexMap[key] = len(m.items) - 1
-	} else {
-		m.items[idx].score = m.items[idx].score.Add(value)
-	}
-
-	m.totalScore = m.totalScore.Add(value)
-}
-
-func (m *scoreMap) Walk(fn func(addr sdk.AccAddress, score sdkmath.Int) error) error {
-	for _, pair := range m.items {
-		if err := fn(pair.addr, pair.score); err != nil {
-			return err
-		}
-	}
 	return nil
 }
